Refactor getMaxAllowedPacketLength in main.go
This commit removes the redundant MaxAllowedPacketMysqlDrv constant and moves the
MaxAllowedPacketLength assignment in the New function. The constant was not necessary as
the maximum packet size is now directly obtained from the database system variable 'max_allowed_packet'
and the assignment was moved to improve the flow and organization of the code. Also, it improves the
closing of the database connection by using defer within an anonymous function to handle any errors.