Add maxAllowedPacket param to New func, remove SetMaxOpenCons
The New function in main.go was updated to include a new parameter, maxAllowedPacket. This allows for the maximum allowed packet size to be set during the creation of a new pool. Additionally, the line that set the maximum open connections was removed to allow the default configuration, as having this set was causing the pool to crash during long server runtime.